refactor: 컨트롤러 테스트에 사용하는 스프링부트 환경을 공유하여 성능 개선
