+++ src/test/java/chess/ControllerTest.java
@@ -1,0 +1,32 @@
+package chess;
+
+import chess.dao.JdbcFixture;
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.test.context.ActiveProfiles;
+
+@ActiveProfiles("test")
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+public abstract class ControllerTest {
+
+    @LocalServerPort
+    protected int port;
+
+    @Autowired
+    protected JdbcTemplate jdbcTemplate;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+        JdbcFixture.dropTable(jdbcTemplate, "square");
+        JdbcFixture.dropTable(jdbcTemplate, "room");
+        JdbcFixture.createRoomTable(jdbcTemplate);
+        JdbcFixture.createSquareTable(jdbcTemplate);
+        JdbcFixture.insertRoom(jdbcTemplate, "roma", "1234", "white");
+    }
+}
